pyproject.toml
[project]
name = "automated-devops-agent"
version = "1.0.0"
description = "Enterprise-grade multi-agent system for automated code review, security scanning, and debugging"
authors = [
    {name = "DevOps Agent Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "google-adk",
    "python-dotenv",
    "pylint>=3.0.0",
    "radon>=6.0.0",
    "pytest>=7.0.0",
    "requests>=2.31.0",
    "googlesearch-python>=1.2.0",
    "PyGithub>=2.1.1",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["automated_devops_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
